Fixes potential issue canceling adding a shot log entry

+++ BeanCounter/BeanCounter/Views/CoffeeEntryPage.xaml.cs
@@ -160,8 +160,11 @@ namespace BeanCounter.Views
 
         private async Task DeleteCoffee()
         {
-            var coffee = ((CoffeeEntryPageViewModel)BindingContext).GenerateCoffee();
-            await App.Database.DeleteCoffeeAsync(coffee);
+            if (ItemId != null)
+            {
+                var coffee = ((CoffeeEntryPageViewModel)BindingContext).GenerateCoffee();
+                await App.Database.DeleteCoffeeAsync(coffee);
+            }
             await NavigateBack();
         }
